+++ django_flexquery/contrib/user_based.py
@@ -1,5 +1,5 @@
 """
-FlexQuery variant that filters the base QuerySet for a given user.
+FlexQuery variant that restricts the base QuerySet for a given user.
 """
 
 from django.http import HttpRequest
@@ -8,23 +8,32 @@ from ..flexquery import FlexQuery
 from ..q import Q
 
 
-__all__ = ["ForUserFlexQuery"]
+__all__ = ["UserBasedFlexQuery"]
 
 
-class ForUserFlexQuery(FlexQuery):
+class UserBasedFlexQuery(FlexQuery):
     """
     This is a slightly modified ``FlexQuery`` implementation, accepting either
     a ``django.http.HttpRequest`` or a user object as argument for the custom
-    function and passing the user through. When no user (or ``None``) is given,
-    the default behavior is to hide all objects. This can be changed by setting the
-    ``all_if_no_user`` attribute to ``True``.
+    function and passing the user through.
+
+    When no user (or ``None``) is given, the behavior is determined by the
+    ``no_user_behavior`` attribute, which may be set to one of the following constants
+    defined on the ``UserBasedFlexQuery`` class:
+
+    * ``NUB_PASS``: let the custom function handle a ``user`` of ``None``
+    * ``NUB_ALL``: don't restrict the queryset
+    * ``NUB_NONE``: restrict to the empty queryset (default)
 
     Because it can handle an ``HttpRequest`` directly, instances of this ``FlexQuery``
     may also be used in conjunction with the django_filters library as the ``queryset``
     parameter of filters.
     """
 
-    all_if_no_user = False
+    NUB_ALL = 0
+    NUB_NONE = 1
+    NUB_PASS = 2
+    no_user_behavior = NUB_NONE
 
     def call_bound(self, user, *args, **kwargs):  # pylint: disable=arguments-differ
         """Calls the custom function with a user, followed by the remaining arguments.
@@ -38,8 +47,9 @@ class ForUserFlexQuery(FlexQuery):
         except AttributeError:
             user = None
         if user is None:
-            if self.all_if_no_user:
+            if self.no_user_behavior == self.NUB_ALL:
                 return Q()
-            return Q(pk__in=self.base.none())
+            if self.no_user_behavior == self.NUB_NONE:
+                return Q(pk__in=self.base.none())
 
         return super().call_bound(user, *args, **kwargs)