+++ tests/testsuite/contrib/test_user_based.py
@@ -1,15 +1,21 @@
 from django.contrib.auth.models import AnonymousUser
 from django.test import RequestFactory, TestCase
 from django_flexquery import *
-from django_flexquery.contrib.for_user import *
+from django_flexquery.contrib.user_based import *
 from django_flexquery_tests.models import *
 
 
-class ForUserFlexQueryTestCase(TestCase):
+class UserIsNone(Exception):
+    pass
+
+
+class UserBasedFlexQueryTestCase(TestCase):
     def setUp(self):
         class Man(Manager):
-            @ForUserFlexQuery.from_func
+            @UserBasedFlexQuery.from_func
             def fq(base, user):
+                if user is None:
+                    raise UserIsNone
                 assert isinstance(user, AnonymousUser)
                 return Q(a=42)
 
@@ -30,9 +36,14 @@ class ForUserFlexQueryTestCase(TestCase):
     def test_request_no_user(self):
         self.assertEqual(self.man.fq(self.req).count(), 0)
 
-    def test_no_user_all(self):
-        self.man.fq.__class__.all_if_no_user = True
+    def test_nub_all(self):
+        self.man.fq.__class__.no_user_behavior = UserBasedFlexQuery.NUB_ALL
         self.assertEqual(self.man.fq(None).count(), 2)
 
-    def test_no_user_none(self):
+    def test_nub_none(self):
         self.assertEqual(self.man.fq(None).count(), 0)
+
+    def test_nub_pass(self):
+        self.man.fq.__class__.no_user_behavior = UserBasedFlexQuery.NUB_PASS
+        with self.assertRaises(UserIsNone):
+            self.man.fq(None)